requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
flask-security-too
flask_restful